Likelihood Estimation

Compute the density of input data.

Usage

lik(arf, params, x, oob = FALSE, log = TRUE, batch = NULL, parallel = TRUE)

Arguments

arf

Pre-trained adversarial random forest. Alternatively, any object of class ranger.

params

Parameters learned via forde.

x

Input data. Function outputs densities for each sample.

oob

Only use out-of-bag leaves for likelihood estimation? If TRUE, x must be the same dataset used to train rf.

log

Return likelihoods on log scale?

batch

Batch size. The default is to compute densities for all of x in one round, which is always the fastest option if memory allows. However, with large samples or many trees, it can be more memory efficient to split the data into batches. This has no impact on results.

parallel

Compute in parallel? Must register backend beforehand, e.g. via doParallel.

Value

A vector of likelihoods, optionally on the log scale.

Details

This function computes the density of input data according to a FORDE model using a pre-trained ARF. Each sample's likelihood is a weighted average of its likelihood in all leaves whose split criteria it satisfies. Intra-leaf densities are fully factorized, since ARFs satisfy the local independence criterion by construction.

References

Watson, D., Blesch, K., Kapar, J., & Wright, M. (2022). Adversarial random forests for density estimation and generative modeling. arXiv preprint, 2205.09435.

See also

Examples

# Estimate average log-likelihood
arf <- adversarial_rf(iris)
#> Iteration: 0, Accuracy: 83.84%
#> Iteration: 1, Accuracy: 37.12%
psi <- forde(arf, iris)
ll <- lik(arf, psi, iris, log = TRUE)
mean(ll)
#> [1] -1.132861