Forests for Density Estimation

Uses a pre-trained ARF model to estimate leaf and distribution parameters.

Usage

forde(
  arf,
  x,
  oob = FALSE,
  family = "truncnorm",
  finite_bounds = c("no", "local", "global"),
  alpha = 0,
  epsilon = 0,
  parallel = TRUE
)

Arguments

arf

Pre-trained adversarial_rf. Alternatively, any object of class ranger.

x

Training data for estimating parameters.

oob

Only use out-of-bag samples for parameter estimation? If TRUE, x must be the same dataset used to train arf.

family

Distribution to use for density estimation of continuous features. Current options include truncated normal (the default family = "truncnorm") and uniform (family = "unif"). See Details.

finite_bounds

Impose finite bounds on all continuous variables? If 'local', infinite bounds are shrinked to empirical extrema within leaves. If 'global', infinite bounds are shrinked to global empirical extrema. if 'no' (the default), do not impose finite bounds.

alpha

Optional pseudocount for Laplace smoothing of categorical features. This avoids zero-mass points when test data fall outside the support of training data. Effectively parametrizes a flat Dirichlet prior on multinomial likelihoods.

epsilon

Optional slack parameter on empirical bounds when finite_bounds != 'no'. This avoids zero-density points when test data fall outside the support of training data. The gap between lower and upper bounds is expanded by a factor of 1 + epsilon.

parallel

Compute in parallel? Must register backend beforehand, e.g. via doParallel.

Value

A list with 5 elements: (1) parameters for continuous data; (2) parameters for discrete data; (3) leaf indices and coverage; (4) metadata on variables; and (5) the data input class. This list is used for estimating likelihoods with lik and generating data with forge.

Details

forde extracts leaf parameters from a pretrained forest and learns distribution parameters for data within each leaf. The former includes coverage (proportion of data falling into the leaf) and split criteria. The latter includes proportions for categorical features and mean/variance for continuous features. The result is a probabilistic circuit, stored as a data.table, which can be used for various downstream inference tasks.

Currently, forde only provides support for a limited number of distributional families: truncated normal or uniform for continuous data, and multinomial for discrete data. Future releases will accommodate a larger set of options.

Though forde was designed to take an adversarial random forest as input, the function's first argument can in principle be any object of class ranger. This allows users to test performance with alternative pipelines (e.g., with supervised forest input). There is also no requirement that x be the data used to fit arf, unless oob = TRUE. In fact, using another dataset here may protect against overfitting. This connects with Wager & Athey's (2018) notion of "honest trees".

References

Watson, D., Blesch, K., Kapar, J., & Wright, M. (2023). Adversarial random forests for density estimation and generative modeling. In Proceedings of the 26th International Conference on Artificial Intelligence and Statistics, pp. 5357-5375.

Wager, S. & Athey, S. (2018). Estimation and inference of heterogeneous treatment effects using random forests. J. Am. Stat. Assoc., 113(523): 1228-1242.

See also

Examples

arf <- adversarial_rf(iris)
#> Iteration: 0, Accuracy: 87.76%
#> Iteration: 1, Accuracy: 36.15%
psi <- forde(arf, iris)
head(psi)
#> $cnt
#> Key: <f_idx, variable>
#>      f_idx     variable   min   max        mu      sigma NA_share
#>      <int>       <char> <num> <num>     <num>      <num>    <num>
#>   1:     1 Petal.Length  -Inf   Inf 5.2000000 0.51478151        0
#>   2:     1  Petal.Width  -Inf  1.70 1.5400000 0.11401754        0
#>   3:     1 Sepal.Length  -Inf   Inf 6.1000000 0.82158384        0
#>   4:     1  Sepal.Width  -Inf   Inf 2.6200000 0.30331502        0
#>   5:     2 Petal.Length  -Inf   Inf 5.6666667 0.30550505        0
#>  ---                                                             
#> 836:   209  Sepal.Width  3.15   Inf 3.2600000 0.08944272        0
#> 837:   210 Petal.Length  -Inf  5.25 2.7500000 1.35757136        0
#> 838:   210  Petal.Width  -Inf  1.20 0.6833333 0.46654760        0
#> 839:   210 Sepal.Length  5.15  5.55 5.4166667 0.11690452        0
#> 840:   210  Sepal.Width  -Inf  3.45 2.9333333 0.51639778        0
#> 
#> $cat
#> Key: <f_idx, variable>
#>      f_idx variable        val      prob NA_share
#>      <int>   <char>     <char>     <num>    <num>
#>   1:     1  Species  virginica 1.0000000        0
#>   2:     2  Species  virginica 1.0000000        0
#>   3:     3  Species  virginica 1.0000000        0
#>   4:     4  Species versicolor 1.0000000        0
#>   5:     5  Species  virginica 1.0000000        0
#>  ---                                             
#> 246:   208  Species     setosa 0.8095238        0
#> 247:   208  Species versicolor 0.1904762        0
#> 248:   209  Species versicolor 1.0000000        0
#> 249:   210  Species     setosa 0.5000000        0
#> 250:   210  Species versicolor 0.5000000        0
#> 
#> $forest
#> Key: <tree, leaf>
#>      f_idx  tree  leaf        cvg
#>      <int> <int> <int>      <num>
#>   1:     1     1    15 0.03333333
#>   2:     2     1    32 0.02000000
#>   3:     3     1    33 0.03333333
#>   4:     4     1    37 0.07333333
#>   5:     5     1    41 0.06666667
#>  ---                             
#> 206:   206    10    77 0.04000000
#> 207:   207    10    78 0.16000000
#> 208:   208    10    80 0.14000000
#> 209:   209    10    83 0.03333333
#> 210:   210    10    85 0.04000000
#> 
#> $meta
#>        variable   class    family decimals
#>          <char>  <char>    <char>    <int>
#> 1: Sepal.Length numeric truncnorm        1
#> 2:  Sepal.Width numeric truncnorm        1
#> 3: Petal.Length numeric truncnorm        1
#> 4:  Petal.Width numeric truncnorm        1
#> 5:      Species  factor  multinom       NA
#> 
#> $levels
#>    variable        val
#>      <fctr>     <char>
#> 1:  Species  virginica
#> 2:  Species     setosa
#> 3:  Species versicolor
#> 
#> $input_class
#> [1] "data.frame"
#>