Forests for Density Estimation

Uses a pre-trained ARF model to estimate leaf and distribution parameters.

Usage

forde(
  arf,
  x,
  oob = FALSE,
  family = "truncnorm",
  alpha = 0,
  epsilon = 0,
  parallel = TRUE
)

Arguments

arf

Pre-trained adversarial_rf. Alternatively, any object of class ranger.

x

Training data for estimating parameters.

oob

Only use out-of-bag samples for parameter estimation? If TRUE, x must be the same dataset used to train arf.

family

Distribution to use for density estimation of continuous features. Current options include truncated normal (the default family = "truncnorm") and uniform (family = "unif"). See Details.

alpha

Optional pseudocount for Laplace smoothing of categorical features. This avoids zero-mass points when test data fall outside the support of training data. Effectively parametrizes a flat Dirichlet prior on multinomial likelihoods.

epsilon

Optional slack parameter on empirical bounds when family = "unif". This avoids zero-density points when test data fall outside the support of training data. The gap between lower and upper bounds is expanded by a factor of 1 + epsilon.

parallel

Compute in parallel? Must register backend beforehand, e.g. via doParallel.

Value

A list with 5 elements: (1) parameters for continuous data; (2) parameters for discrete data; (3) leaf indices and coverage; (4) metadata on variables; and (5) the data input class. This list is used for estimating likelihoods with lik and generating data with forge.

Details

forde extracts leaf parameters from a pretrained forest and learns distribution parameters for data within each leaf. The former includes coverage (proportion of data falling into the leaf) and split criteria. The latter includes proportions for categorical features and mean/variance for continuous features. These values are stored in a data.table, which can be used as input to various other functions.

Currently, forde only provides support for a limited number of distributional families: truncated normal or uniform for continuous data, and multinomial for discrete data. Future releases will accommodate a larger set of options.

Though forde was designed to take an adversarial random forest as input, the function's first argument can in principle be any object of class ranger. This allows users to test performance with alternative pipelines (e.g., with supervised forest input). There is also no requirement that x be the data used to fit arf, unless oob = TRUE. In fact, using another dataset here may protect against overfitting. This connects with Wager & Athey's (2018) notion of "honest trees".

References

Watson, D., Blesch, K., Kapar, J., & Wright, M. (2022). Adversarial random forests for density estimation and generative modeling. arXiv preprint, 2205.09435.

Wager, S. & Athey, S. (2018). Estimation and inference of heterogeneous treatment effects using random forests. J. Am. Stat. Assoc., 113(523): 1228-1242.

See also

Examples

arf <- adversarial_rf(iris)
#> Iteration: 0, Accuracy: 87.58%
#> Iteration: 1, Accuracy: 44.33%
psi <- forde(arf, iris)
head(psi)
#> $cnt
#>      f_idx     variable  min  max       mu       sigma
#>   1:     1 Petal.Length -Inf  Inf 5.933333 0.208166600
#>   2:     1  Petal.Width 2.45  Inf 2.500000 0.019937845
#>   3:     1 Sepal.Length -Inf  Inf 6.733333 0.450924975
#>   4:     1  Sepal.Width -Inf  Inf 3.400000 0.173205081
#>   5:     2 Petal.Length -Inf 5.05 1.406250 0.211246302
#>  ---                                                  
#> 660:   165  Sepal.Width 2.85 3.25 3.120000 0.109544512
#> 661:   166 Petal.Length -Inf 1.35 1.150000 0.212132034
#> 662:   166  Petal.Width -Inf 0.25 0.200000 0.009886127
#> 663:   166 Sepal.Length -Inf 5.60 5.050000 0.636396103
#> 664:   166  Sepal.Width 3.25  Inf 3.550000 0.070710678
#> 
#> $cat
#>      f_idx variable        val      prob
#>   1:     1  Species  virginica 1.0000000
#>   2:     2  Species     setosa 1.0000000
#>   3:     3  Species versicolor 0.2500000
#>   4:     3  Species  virginica 0.7500000
#>   5:     4  Species versicolor 0.1666667
#>  ---                                    
#> 193:   162  Species     setosa 1.0000000
#> 194:   163  Species     setosa 1.0000000
#> 195:   164  Species versicolor 1.0000000
#> 196:   165  Species     setosa 1.0000000
#> 197:   166  Species     setosa 1.0000000
#> 
#> $forest
#>      f_idx tree leaf        cvg
#>   1:     1    1    3 0.02000000
#>   2:     2    1   18 0.10666667
#>   3:     3    1   24 0.02666667
#>   4:     4    1   38 0.04000000
#>   5:     5    1   44 0.05333333
#>  ---                           
#> 162:   162   10   61 0.02000000
#> 163:   163   10   65 0.01333333
#> 164:   164   10   66 0.03333333
#> 165:   165   10   68 0.03333333
#> 166:   166   10   69 0.01333333
#> 
#> $meta
#>        variable   class    family
#> 1: Sepal.Length numeric truncnorm
#> 2:  Sepal.Width numeric truncnorm
#> 3: Petal.Length numeric truncnorm
#> 4:  Petal.Width numeric truncnorm
#> 5:      Species  factor  multinom
#> 
#> $input_class
#> [1] "data.frame"
#>