Forests for Density Estimation

Uses a pre-trained ARF model to estimate leaf and distribution parameters.

Usage

forde(
  arf,
  x,
  oob = FALSE,
  family = "truncnorm",
  epsilon = 0.1,
  parallel = TRUE
)

Arguments

arf

Pre-trained adversarial_rf. Alternatively, any object of class ranger.

x

Training data for estimating parameters.

oob

Only use out-of-bag samples for parameter estimation? If TRUE, x must be the same dataset used to train arf.

family

Distribution to use for density estimation of continuous features. Current options include truncated normal (the default family = "truncnorm") and uniform (family = "unif"). See Details.

epsilon

Slack parameter on empirical bounds when family = "unif". This avoids zero-density points when test data fall outside the support of training data. The gap between lower and upper bounds is expanded by a factor of epsilon. Only used when a variable is never selected for splitting.

parallel

Compute in parallel? Must register backend beforehand, e.g. via doParallel.

Value

A list with 4 elements: (1) parameters for continuous data; (2) parameters for discrete data; (3) leaf indices and coverage; and (4) metadata on variables. This list is used for estimating likelihoods with lik and generating data with forge.

Details

forde extracts leaf parameters from a pretrained forest and learns distribution parameters for data within each leaf. The former includes coverage (proportion of data falling into the leaf) and split criteria. The latter includes proportions for categorical features and mean/variance for continuous features. These values are stored in a data.table, which can be used as input to various other functions.

Currently, forde only provides support for a limited number of distributional families: truncated normal or uniform for continuous data, and multinomial for discrete data. Future releases will accommodate a larger set of options.

Though forde was designed to take an adversarial random forest as input, the function's first argument can in principle be any object of class ranger. This allows users to test performance with alternative pipelines (e.g., with supervised forest input). There is also no requirement that x be the data used to fit arf, unless oob = TRUE. In fact, using another dataset here may protect against overfitting. This connects with Wager & Athey's (2018) notion of "honest trees".

References

Watson, D., Blesch, K., Kapar, J., & Wright, M. (2022). Adversarial random forests for density estimation and generative modeling. arXiv preprint, 2205.09435.

Wager, S. & Athey, S. (2018). Estimation and inference of heterogeneous treatment effects using random forests. J. Am. Stat. Assoc., 113(523): 1228-1242.

See also

Examples

arf <- adversarial_rf(iris)
#> Iteration: 0, Accuracy: 84.69%
#> Iteration: 1, Accuracy: 42.57%
psi <- forde(arf, iris)
head(psi)
#> $cnt
#>          variable  min  max f_idx       mu        sigma
#>   1: Sepal.Length -Inf  Inf    12 5.155556 0.2172083957
#>   2: Sepal.Length -Inf  Inf    21 4.900000 0.1000000000
#>   3: Sepal.Length -Inf  Inf     4 4.511111 0.1452966315
#>   4: Sepal.Length -Inf 4.75     2 4.600000 0.0008211816
#>   5: Sepal.Length -Inf  Inf    18 5.200000 0.6745368782
#>  ---                                                   
#> 676:  Petal.Width -Inf  Inf   168 1.475000 0.0500000000
#> 677:  Petal.Width -Inf  Inf   162 1.862500 0.0916125381
#> 678:  Petal.Width -Inf 2.35   155 2.009091 0.2553963052
#> 679:  Petal.Width -Inf  Inf   158 2.460000 0.0547722558
#> 680:  Petal.Width -Inf  Inf   154 2.000000 0.2708012802
#> 
#> $cat
#>      variable       val       prob f_idx
#>   1:  Species    setosa 1.00000000    12
#>   2:  Species    setosa 1.00000000    21
#>   3:  Species    setosa 1.00000000     4
#>   4:  Species    setosa 1.00000000     2
#>   5:  Species    setosa 0.80000000    18
#>  ---                                    
#> 198:  Species virginica 1.00000000   154
#> 199:  Species virginica 0.96969697   155
#> 200:  Species virginica 0.09090909   164
#> 201:  Species virginica 0.87500000   162
#> 202:  Species virginica 0.14285714   165
#> 
#> $forest
#>      f_idx tree leaf        cvg
#>   1:     1    1   12 0.02000000
#>   2:     2    1   17 0.01333333
#>   3:     3    1   26 0.04000000
#>   4:     4    1   29 0.06000000
#>   5:     5    1   38 0.04666667
#>  ---                           
#> 166:   166   10   68 0.16000000
#> 167:   167   10   74 0.08000000
#> 168:   168   10   79 0.02666667
#> 169:   169   10   82 0.08666667
#> 170:   170   10   85 0.01333333
#> 
#> $meta
#>        variable   class    family
#> 1: Sepal.Length numeric truncnorm
#> 2:  Sepal.Width numeric truncnorm
#> 3: Petal.Length numeric truncnorm
#> 4:  Petal.Width numeric truncnorm
#> 5:      Species  factor  multinom
#>