Forests for Density Estimation

Uses a pre-trained ARF model to estimate leaf and distribution parameters.

Usage

forde(
  arf,
  x,
  oob = FALSE,
  family = "truncnorm",
  alpha = 0,
  epsilon = 0,
  parallel = TRUE
)

Arguments

arf

Pre-trained adversarial_rf. Alternatively, any object of class ranger.

x

Training data for estimating parameters.

oob

Only use out-of-bag samples for parameter estimation? If TRUE, x must be the same dataset used to train arf.

family

Distribution to use for density estimation of continuous features. Current options include truncated normal (the default family = "truncnorm") and uniform (family = "unif"). See Details.

alpha

Optional pseudocount for Laplace smoothing of multinomial features. This avoids zero-mass points when test data fall outside the support of training data. Effectively parametrizes a flat Dirichlet prior on multinomial likelihoods.

epsilon

Optional slack parameter on empirical bounds when family = "unif". This avoids zero-density points when test data fall outside the support of training data. The gap between lower and upper bounds is expanded by a factor of 1 + epsilon.

parallel

Compute in parallel? Must register backend beforehand, e.g. via doParallel.

Value

A list with 5 elements: (1) parameters for continuous data; (2) parameters for discrete data; (3) leaf indices and coverage; (4) metadata on variables; and (5) the data input class. This list is used for estimating likelihoods with lik and generating data with forge.

Details

forde extracts leaf parameters from a pretrained forest and learns distribution parameters for data within each leaf. The former includes coverage (proportion of data falling into the leaf) and split criteria. The latter includes proportions for categorical features and mean/variance for continuous features. These values are stored in a data.table, which can be used as input to various other functions.

Currently, forde only provides support for a limited number of distributional families: truncated normal or uniform for continuous data, and multinomial for discrete data. Future releases will accommodate a larger set of options.

Though forde was designed to take an adversarial random forest as input, the function's first argument can in principle be any object of class ranger. This allows users to test performance with alternative pipelines (e.g., with supervised forest input). There is also no requirement that x be the data used to fit arf, unless oob = TRUE. In fact, using another dataset here may protect against overfitting. This connects with Wager & Athey's (2018) notion of "honest trees".

References

Watson, D., Blesch, K., Kapar, J., & Wright, M. (2022). Adversarial random forests for density estimation and generative modeling. arXiv preprint, 2205.09435.

Wager, S. & Athey, S. (2018). Estimation and inference of heterogeneous treatment effects using random forests. J. Am. Stat. Assoc., 113(523): 1228-1242.

See also

Examples

arf <- adversarial_rf(iris)
#> Iteration: 0, Accuracy: 88.26%
#> Iteration: 1, Accuracy: 45.27%
psi <- forde(arf, iris)
head(psi)
#> $cnt
#>      f_idx     variable  min  max        mu      sigma
#>   1:     1 Petal.Length -Inf  Inf 1.4666667 0.05773503
#>   2:     1  Petal.Width -Inf 2.35 0.2333333 0.15275252
#>   3:     1 Sepal.Length -Inf  Inf 5.4666667 0.25166115
#>   4:     1  Sepal.Width 4.05  Inf 4.2333333 0.15275252
#>   5:     2 Petal.Length -Inf  Inf 5.6833333 0.35449495
#>  ---                                                  
#> 692:   173  Sepal.Width 2.95 3.30 3.0625000 0.11877349
#> 693:   174 Petal.Length -Inf 1.50 1.2800000 0.10954451
#> 694:   174  Petal.Width -Inf  Inf 0.2000000 0.07071068
#> 695:   174 Sepal.Length -Inf 4.55 4.4000000 0.07071068
#> 696:   174  Sepal.Width -Inf  Inf 2.8800000 0.34205263
#> 
#> $cat
#>      f_idx variable        val  prob
#>   1:     1  Species     setosa 1.000
#>   2:     2  Species  virginica 1.000
#>   3:     3  Species     setosa 1.000
#>   4:     4  Species     setosa 1.000
#>   5:     5  Species  virginica 1.000
#>  ---                                
#> 192:   171  Species     setosa 1.000
#> 193:   172  Species versicolor 1.000
#> 194:   173  Species     setosa 0.125
#> 195:   173  Species versicolor 0.875
#> 196:   174  Species     setosa 1.000
#> 
#> $forest
#>      f_idx tree leaf        cvg
#>   1:     1    1    5 0.02000000
#>   2:     2    1    7 0.04000000
#>   3:     3    1   15 0.02666667
#>   4:     4    1   16 0.02000000
#>   5:     5    1   46 0.02000000
#>  ---                           
#> 170:   170   10   67 0.01333333
#> 171:   171   10   70 0.04000000
#> 172:   172   10   77 0.14000000
#> 173:   173   10   79 0.05333333
#> 174:   174   10   80 0.03333333
#> 
#> $meta
#>        variable   class    family
#> 1: Sepal.Length numeric truncnorm
#> 2:  Sepal.Width numeric truncnorm
#> 3: Petal.Length numeric truncnorm
#> 4:  Petal.Width numeric truncnorm
#> 5:      Species  factor  multinom
#> 
#> $input_class
#> [1] "data.frame"
#>