Forests for Density Estimation

Uses a pre-trained ARF model to estimate leaf and distribution parameters.

Usage

forde(
  arf,
  x,
  oob = FALSE,
  family = "truncnorm",
  epsilon = 0.1,
  parallel = TRUE
)

Arguments

arf

Pre-trained adversarial random forest. Alternatively, any object of class ranger.

x

Training data for estimating parameters.

oob

Only use out-of-bag samples for parameter estimation? If TRUE, x must be the same dataset used to train arf.

family

Distribution to use for density estimation of continuous features. Current options include truncated normal (the default family = "truncnorm") and uniform (family = "unif"). See Details.

epsilon

Slack parameter on empirical bounds when family = "unif". This avoids zero-density points when test data fall outside the support of training data. The gap between lower and upper bounds is expanded by a factor of epsilon. Only used when a variable is never selected for splitting.

parallel

Compute in parallel? Must register backend beforehand, e.g. via doParallel.

Value

A data.table with leaf and distribution parameters for each feature. These are used for estimating likelihoods with lik and generating data with forge.

Details

forde extracts leaf parameters from a pretrained forest and learns distribution parameters for data within each leaf. The former includes coverage (proportion of data falling into the leaf) and split criteria. The latter includes proportions for categorical features and mean/variance for continuous features. These values are stored in a data.table called psi, which can be passed onto forge for data synthesis. They are also used to compute log-likelihoods for x_trn or x_tst if loglik = TRUE.

Currently, forde only provides support for truncated normal or uniform densities when features are continuous. Future releases will accommodate a larger class of distributional families.

Though forde was designed to take an adversarial random forest as input, arf can in principle be any object of class ranger. This allows users to test performance with alternative pipelines (e.g., with supervised forest input). There is also no requirement that x_trn be the data used to fit arf.

References

Watson, D., Blesch, K., Kapar, J., & Wright, M. (2022). Adversarial random forests for density estimation and generative modeling. arXiv preprint, 2205.09435.

See also

Examples

arf <- adversarial_rf(iris)
#> Iteration: 0, Accuracy: 87.84%
#> Iteration: 1, Accuracy: 39.39%
psi <- forde(arf, iris)
head(psi)
#>    tree leaf     variable       cat prob       mu     sigma    family  min  max
#> 1:    1   24 Petal.Length      <NA>   NA 6.190909 0.3645670 truncnorm -Inf  Inf
#> 2:    1   24  Petal.Width      <NA>   NA 1.954545 0.1916436 truncnorm -Inf 2.25
#> 3:    1   24 Sepal.Length      <NA>   NA 7.300000 0.4289522 truncnorm -Inf  Inf
#> 4:    1   24  Sepal.Width      <NA>   NA 3.072727 0.4002272 truncnorm -Inf  Inf
#> 5:    1   24      Species virginica    1       NA        NA  multinom -Inf  Inf
#> 6:    1   29 Petal.Length      <NA>   NA 5.450000 0.1643168 truncnorm -Inf  Inf
#>           cvg
#> 1: 0.07333333
#> 2: 0.07333333
#> 3: 0.07333333
#> 4: 0.07333333
#> 5: 0.07333333
#> 6: 0.04000000