Forests for Generative Modeling

Uses pre-trained FORDE model to simulate synthetic data.

Usage

forge(params, n_synth, parallel = TRUE)

Arguments

params

Parameters learned via forde.

n_synth

Number of synthetic samples to generate.

parallel

Compute in parallel? Must register backend beforehand, e.g. via doParallel.

Value

A dataset of n_synth synthetic samples.

Details

forge simulates a synthetic dataset of n_synth samples. First, leaves are sampled in proportion to their coverage. Then, each feature is sampled independently within each leaf according to the probability mass or density function learned by forde. This will create realistic data so long as the adversarial RF used in the previous step satisfies the local independence criterion. See Watson et al. (2022).

References

Watson, D., Blesch, K., Kapar, J., & Wright, M. (2022). Adversarial random forests for density estimation and generative modeling. arXiv preprint, 2205.09435.

See also

Examples

arf <- adversarial_rf(iris)
#> Iteration: 0, Accuracy: 87.16%
#> Iteration: 1, Accuracy: 36.12%
psi <- forde(arf, iris)
x_synth <- forge(psi, n_synth = 100)